latest

+++ app/src/main/java/com/example/tapandrepair/MechanicData.kt
@@ -4,5 +4,5 @@ data class MechanicData(
     val acceptance: Int,
     val cancellation: Int,
     val mechanic: MechanicXX,
-    val rating: Int
+    val rating: Double
 )
